// libs/proposal/src/types/proposal-policy-label.ts
export enum ProposalPolicyLabel {
  Config = 'config',
  Policy = 'policy',
  AddMemberToRole = 'add_member_to_role',
  RemoveMemberFromRole = 'remove_member_from_role',
  Call = 'call',
  UpgradeSelf = 'upgrade_self',
  UpgradeRemote = 'upgrade_remote',
  Transfer = 'transfer',
  SetVoteToken = 'set_vote_token',
  AddBounty = 'add_bounty',
  BountyDone = 'bounty_done',
  Vote = 'vote',
}